mysqlp/app/requirements.txt
flask
mysql-connector
boto3